+++ programs/encore/src/instructions/mod.rs
@@ -1,5 +1,7 @@
 pub mod event_create;
 pub mod event_update;
+pub mod ticket_mint;
 
 pub use event_create::*;
 pub use event_update::*;
+pub use ticket_mint::*;