+++ programs/encore/src/state/mod.rs
@@ -1,5 +1,7 @@
 //! State account definitions
 
 pub mod event_config;
+pub mod ticket;
 
 pub use event_config::*;
+pub use ticket::*;